refactor: address code review feedback for final reports module
## Architecture (Critical)
- Split service file (482 lines) into subdomain files:
  - disciplina-relatorio-service.ts (~220 lines)
  - monitor-relatorio-service.ts (~280 lines)
  - utils.ts (safe JSON parsing helper)
  - relatorios-finais-service.ts (facade, ~40 lines)

## Security (Critical)
- Fix JSON parsing to use Zod schema validation
- Add safeJsonParse<T>(json, schema, context) helper with runtime validation

## Performance
- Add composite index for pending signature queries:
  - relatorio_monitor_pending_signature_idx (status, alunoAssinouEm, inscricaoId)

## Frontend
- Extract components from 618-line page:
  - ProjetoCard.tsx
  - ProjetoDetailsDialog.tsx
  - RelatorioDisciplinaDialog.tsx
  - RelatorioMonitorDialog.tsx
  - RelatorioStatusBadge.tsx
- Main page now ~300 lines
- Extract YEARS_TO_SHOW constant
- Add JSDoc comments to service methods
