+++ src/app/home/professor/relatorios-finais/components/index.ts
@@ -1,0 +1,5 @@
+export { ProjetoCard } from './ProjetoCard'
+export { ProjetoDetailsDialog } from './ProjetoDetailsDialog'
+export { RelatorioDisciplinaDialog } from './RelatorioDisciplinaDialog'
+export { RelatorioMonitorDialog } from './RelatorioMonitorDialog'
+export { RelatorioStatusBadge } from './RelatorioStatusBadge'
